requirements.txt
openai>=2.16.0,<3.0.0
numpy
Pillow